// frontend/src/api/models/handlerProductResponse.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */
import type { HandlerProductResponseStatus } from './handlerProductResponseStatus'

/**
 * Product details returned by the API
 */
export interface HandlerProductResponse {
  attributes?: string
  barcode?: string
  category_id?: string
  code?: string
  created_at?: string
  description?: string
  id?: string
  min_stock?: number
  name?: string
  profit_margin?: number
  purchase_price?: number
  selling_price?: number
  sort_order?: number
  status?: HandlerProductResponseStatus
  tenant_id?: string
  unit?: string
  updated_at?: string
  version?: number
}