// frontend/src/api/models/handlerProductResponseStatus.ts
/**
 * // This file is auto-generated by orval
 * // Do not edit manually
 *
 */

export type HandlerProductResponseStatus =
  (typeof HandlerProductResponseStatus)[keyof typeof HandlerProductResponseStatus]

export const HandlerProductResponseStatus = {
  active: 'active',
  inactive: 'inactive',
  discontinued: 'discontinued',
} as const